+++ src/com/nonolite/design/Design.java
@@ -3,15 +3,15 @@ package com.nonolite.design;
 public interface Design {
     void background();
     
-    void base(int x, int y, int width, int height);
+    void base(float x, float y, float width, float height);
     
-    void baseRect(int x, int y, int width, int height);
+    void baseRect(float x, float y, float width, float height);
     
-    void baseRect2(int x, int y, int width, int height);
+    void baseRect2(float x, float y, float width, float height);
     
-    void rect1(int x, int y, int width, int height);
+    void rect1(float x, float y, float width, float height);
     
-    void rect2(int x, int y, int width, int height);
+    void rect2(float x, float y, float width, float height);
     
     void text(char text, float x, float y, float width, float height);
     
@@ -31,7 +31,7 @@ public interface Design {
     
     void text(String text, float x, float y, float size, int align1, int align2);
     
-    void symbol1(int x, int y, int width, int height);
+    void symbol1(float x, float y, float width, float height);
     
-    void symbol2(int x, int y, int width, int height);
+    void symbol2(float x, float y, float width, float height);
 }
