+++ src/com/nonolite/layouts/utils/OnClickListener.java
@@ -1,5 +1,5 @@
 package com.nonolite.layouts.utils;
 
 public interface OnClickListener {
-    String onClick(int keyCode, int x, int y);
+    String onClick(int keyCode, float x, float y);
 }
